Refactor workspace package import workflows
