+++ apps/backend/src/workspacePackages/import/apply/importCards.ts
@@ -4,14 +4,14 @@ import {
   type Card,
   type CardMutationMetadata,
   type CreateCardInput,
-} from "../../cards";
+} from "../../../cards";
 import {
   transactionWithWorkspaceScope,
   type DatabaseExecutor,
   type WorkspaceDatabaseScope,
-} from "../../database";
-import { HttpError } from "../../shared/errors";
-import type { WorkspacePackageImportPlannedCard } from "./importPlan";
+} from "../../../database";
+import { HttpError } from "../../../shared/errors";
+import type { WorkspacePackageImportPlannedCard } from "../planning/importPlan";
 
 const workspacePackageImportCardPersistenceBatchSize = 100;
