+++ apps/backend/src/workspacePackages/import/apply/importConfirm.test.ts
@@ -1,13 +1,13 @@
 import assert from "node:assert/strict";
 import { Buffer } from "node:buffer";
 import test from "node:test";
-import type { Card } from "../../cards";
+import type { Card } from "../../../cards";
 import {
   imageJpegCardMediaBlobMimeType,
   type MediaAsset,
-} from "../../mediaAssets/types";
-import type { BackendObservationScope } from "../../observability/sentry";
-import { HttpError } from "../../shared/errors";
+} from "../../../mediaAssets/types";
+import type { BackendObservationScope } from "../../../observability/sentry";
+import { HttpError } from "../../../shared/errors";
 import {
   confirmWorkspacePackageImportWithDependencies,
   type WorkspacePackageImportConfirmDependencies,
@@ -23,8 +23,8 @@ import {
   type WorkspacePackageImportPlanInput,
   type WorkspacePackageImportReferencedMediaFile,
   type WorkspacePackageImportReferencedMediaInput,
-} from "../index";
-import { validateWorkspacePackageImportPlanPreflight } from "./importPlan";
+} from "../../index";
+import { validateWorkspacePackageImportPlanPreflight } from "../planning/importPlan";
 
 const testUserId = "user-1";
 const testWorkspaceId = "11111111-1111-4111-8111-111111111111";
