+++ apps/backend/src/workspacePackages/import/media/importMedia.test.ts
@@ -1,7 +1,7 @@
 import assert from "node:assert/strict";
 import { Buffer } from "node:buffer";
 import test from "node:test";
-import { HttpError } from "../../shared/errors";
+import { HttpError } from "../../../shared/errors";
 import {
   loadWorkspacePackageImportReferencedMedia,
   loadWorkspacePackageImportReferencedMediaWithLimits,
@@ -10,11 +10,11 @@ import {
   type WorkspacePackageCardsJsonV1,
   type WorkspacePackageImportReferencedMediaInput,
   type WorkspacePackageImportReferencedMediaLimits,
-} from "../index";
+} from "../../index";
 import {
   createPackageZip,
   createStoredZip,
-} from "./testZipHelpers";
+} from "../testZipHelpers";
 
 const testMetadata: WorkspacePackageCardMetadataV1 = {
   version: 1,
