+++ apps/backend/src/workspacePackages/import/media/importMedia.ts
@@ -13,8 +13,8 @@ import {
   readWorkspacePackageZipEntryBytes,
   type CollectedWorkspacePackageZip,
   type WorkspacePackageImportZipLimits,
-} from "./importZip";
-import type { WorkspacePackageCardsJsonV1 } from "../types";
+} from "../importZip";
+import type { WorkspacePackageCardsJsonV1 } from "../../types";
 
 export type WorkspacePackageImportReferencedMediaInput = Readonly<{
   packageBytes: Buffer | Uint8Array;
