+++ apps/backend/src/workspacePackages/import/planning/importPreview.test.ts
@@ -1,7 +1,7 @@
 import assert from "node:assert/strict";
 import { Buffer } from "node:buffer";
 import test from "node:test";
-import { HttpError } from "../../shared/errors";
+import { HttpError } from "../../../shared/errors";
 import {
   buildSuggestedWorkspacePackageImportTag,
   createDefaultWorkspacePackageImportTagPolicy,
@@ -12,13 +12,13 @@ import {
   type WorkspacePackageCardMetadataV1,
   type WorkspacePackageCardsJsonV1,
   type WorkspacePackageImportPreviewInput,
-} from "../index";
+} from "../../index";
 import {
   createCardsJsonBuffer,
   createDeflatedZipEntry,
   createPackageZip,
   createStoredZip,
-} from "./testZipHelpers";
+} from "../testZipHelpers";
 
 const generatedAt = "2026-06-30T12:00:00.000Z";
